ASoC: msm: qdsp6: add support for any number of ASM buffers
Current logic in ASM driver supports CPU and DSP buffers only
when the number of buffers is power of 2. At times, the number
of buffers might need to be different than power of 2. Modify
the logic to find next available CPU or DSP buffer, even when
the number of buffers is anything other than power of 2.

Change-Id: I5632d14996fd88906ed3e202bd30adbb45b5b32a
CRs-Fixed: 771446
Signed-off-by: Banajit Goswami <bgoswami@codeaurora.org>