perf: duplicate deletion of perf event
a malicious app can open a perf event with constraint_duplicate
bit set, disable the event, and close the fd.  On closing the fd,
the perf_release() modification causes the kernel to clean up
the event as if it still were enabled, leading to the event
being removed from a list twice.

Change-Id: I11e2a308a110fff224c0b362d6020c04ec05c4c5
CRs-Fixed: 977563
Signed-off-by: Srinivasarao P <spathi@codeaurora.org>